+++ Banks/Entities/Transaction.cs
@@ -4,14 +4,14 @@ namespace Banks.Entities
 {
     public class Transaction
     {
-        public Transaction(string name, float sum, List<Account> accounts)
+        public Transaction(string name, double sum, List<Account> accounts)
         {
             BankAccounts = accounts;
             Sum = sum;
             TranactionName = name;
         }
 
-        public float Sum { get; }
+        public double Sum { get; }
         public string TranactionName { get; }
         public List<Account> BankAccounts { get; }
